# Conflicts:
#	backend_core/users/templates/account/signup.html

+++ backend_core/photos/views.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-from django.shortcuts import render
 from django.http import JsonResponse
 from django.views import View
 from django.conf import settings
@@ -11,6 +10,9 @@ from django.contrib.contenttypes.models import ContentType
 import datetime
 import pytz
 import os
+from django.http import HttpResponseNotFound
+from django.shortcuts import render,redirect
+
 # Create your views here.
 class BasicUploadView(View):
     def get(self, request):
@@ -19,45 +21,65 @@ class BasicUploadView(View):
 
     def post(self, request):
         form = PhotoForm(self.request.POST, self.request.FILES)
-        #print('test1')
-        #print(vars(form))
-        #print(form.errors)
+        # print('test1')
+        # print(vars(form))
+        # print(form.errors)
 
         if form.is_valid():
             f = form.save(commit=False)
             contractor = Contractor.objects.get(pk='1025362')
             f.content_type = ContentType.objects.get(model='contractor')
             f.object_id = contractor.LicNum
-            #print('test2')
+            # print('test2')
             f.save()
             data = {'is_valid': True, 'name': f.img.name, 'url': f.img.url}
         else:
-            #print('test3')
+            # print('test3')
             data = {'is_valid': False}
         return JsonResponse(data)
 
-#@login_required
-def background_photo_upload(request,contractor_id):
-    template_name = 'photos/background_photo_upload.html'
-    success_url = "disk/uploadsuccess.html"
-    if request.method == 'POST':
-        form = BackgroundPhotoForm(request.POST, request.FILES)
-        if form.is_valid():
-            bp, created = BackgroundPhoto.objects.get_or_create(content_type=ContentType.objects.get(model='contractor'),
-                                                                object_id=contractor_id)
 
-            if not created:
-                old_pic_path = bp.img.file.name
+# @login_required
+def background_photo_upload(request, o_id):
+    if request.user.is_anonymous():
+        p_lic_num = None
+    else:
+        try:
+            p_lic_num = int(request.user.professional_profiles.first().professional.lic_num)
+        except:
+            p_lic_num = None
+
+    if str(p_lic_num) == str(o_id):
+        template_name = 'photos/background_photo_upload.html'
+
+        if request.method == 'POST':
+            if 'contractor' in request.path:
+                model_type = 'contractor'
+            elif 'designer' in request.path:
+                model_type = 'designer'
+            elif 'architect' in request.path:
+                model_type = 'architect'
+            success_url = "/"+model_type+"/"+o_id
+            form = BackgroundPhotoForm(request.POST, request.FILES)
+            if form.is_valid():
+                bp, nonexist = BackgroundPhoto.objects.get_or_create(content_type=ContentType.objects.get(model=model_type),
+                                                                    object_id=o_id)
+
+                if not nonexist:
+                    old_pic_path = bp.img.file.name
+                    os.remove(old_pic_path)
 
                 bp.img = form.cleaned_data.get('img')
                 bp.title = form.cleaned_data.get('img').name
                 bp.uploaded_at = datetime.datetime.now(pytz.timezone('UTC'))
-            bp.save()
-            os.remove(old_pic_path)
-            return render(request, success_url)
+                bp.save()
+
+                return redirect(success_url)
+        else:
+            form = BackgroundPhotoForm()
+        return render(request, template_name, {'form': form})
     else:
-        form = BackgroundPhotoForm()
-    return render(request, template_name, {'form': form})
+        return HttpResponseNotFound('No Pages Found.')
 
 def FileFieldUpload(request):
     template_name = 'photos/upload_multiple_files.html'  # Replace with your template.
@@ -68,7 +90,9 @@ def FileFieldUpload(request):
         files = request.FILES.getlist('img')
         if form.is_valid():
             for f in files:
-                instance = FileField.objects.create(img=f, title=f.name, object_id=18, content_type=ContentType.objects.get_for_model(settings.AUTH_USER_MODEL))
+                instance = FileField.objects.create(img=f, title=f.name, object_id=18,
+                                                    content_type=ContentType.objects.get_for_model(
+                                                        settings.AUTH_USER_MODEL))
                 instance.save()
             return render(request, success_url)
     form = FileFieldForm()
@@ -95,4 +119,3 @@ def FileFieldUpload(request):
 #         #else:
 #         #    return self.form_invalid(form)
 #     return render(request, template_name)
-
